requirements.txt
torch>=1.8
torchvision
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git
Pillow